LPS-161841 We are iterating over the articles backwards, so we must start from the last version that we need and decrement it by 0.1 each time.
This is the correct solution for this issue. WE MUST NOT SOLVE THIS BY
ITERATING OVER THE ARTICLES FORWARDS! If we attempt to implement that
"solution", then there is still the possibility of a IX_D3ACAD4A
violation, since the earlier version may get updated to the next version
before the next version gets updated to the version after that.